Implement some optimizations for NEON
We now use as much information as we can (bloom filter results, hangul
filter results) to selectively decompose target code points, and ignore
irrelevant ones, as opposed to falling back on the scalar
implementation. This provides a noticeable speedup to the `korean.txt`
benchmark, and minor speedups for other benchmarks.

This commit also changes the build system to use the amalgamated C file
for release builds.