Fix amalgamation
Previously, the amgalgamation would include C sources files from the
original source tree (defeating the purpose of the amalgmation) as part
of the compile time parameterization scheme built for the compatibility
variants of the standard normalization algorithms.

This commit creates an alternate compile-time parameterization system
that using scripts that are run at build-time.