Drastically speed up decomposition lookups
We now use a trie for looking up BMP code points. This removes the need
for a bloom filter and a minimal perfect hash table. On `odyessy.txt`,
we get a 33% speed up. We get a 5-10% speed up on most other NEON NF(K)D
UTF-8 benchmarks. Right now this is only implemented for UTF-8.

The scalar UTF-8 implementations of decomposition also get a nice speed
up.