Only sort combining characters in NF(K)D when necessary
This leads to a general speed improvement since we no longer check if we
have a sequence <x, y> where x is a combining character and y is a
starter. Most notably, though, we get a very large speedup in
`odyessy_nfd.txt`, since we no longer call the sort function when
combining characters are already ordered.